+++ src/community/components/table/table.stories.tsx
@@ -1,5 +1,5 @@
 import { faker } from '@faker-js/faker';
-import { Meta, StoryFn, StoryObj } from '@storybook/react';
+import { Meta, StoryFn, StoryObj } from '@storybook/react-vite';
 import {
   ColumnDef,
   ColumnFiltersState,